Implement delay, delay_ms, and delay_us
Refactor delay_ms and delay_us to prevent overflow
.
Change delay_ms and delay_us to call the internal delay function with
a 48 bit counter. This prevents delay_ms and delay_us from overflowing
when passed a large argument.
.
Move the internal delay functions to their own module. Remove the
unused internal delay_loop_3_cycles function.
